# docs/swagger.yaml
basePath: /
definitions:
  accountsapi.LoginLocalReq:
    properties:
      login:
        example: jdoe
        type: string
      password:
        example: secret123
        type: string
    type: object
  accountsapi.RegisterLocalReq:
    properties:
      first_name:
        example: John
        type: string
      grant_can_create:
        example: true
        type: boolean
      group_number:
        example: CS-101
        type: string
      last_name:
        example: Doe
        type: string
      login:
        example: jdoe
        type: string
      middle_name:
        example: Q
        type: string
      password:
        example: secret123
        type: string
    type: object
  domain.Account:
    properties:
      createdAt:
        type: string
      firstName:
        type: string
      groupNumber:
        type: string
      id:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: Research Workflow Management System
  title: RWMS API
  version: "0.1"
paths:
  /api/accounts/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - accounts
  /api/auth/login-local:
    post:
      consumes:
      - application/json
      parameters:
      - description: login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accountsapi.LoginLocalReq'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with local credentials
      tags:
      - auth
  /api/auth/register-local:
    post:
      consumes:
      - application/json
      description: Create a new account with username/password
      parameters:
      - description: registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accountsapi.RegisterLocalReq'
      produces:
      - application/json
      responses:
        "200":
          description: account_id
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a local account
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Use "Bearer <JWT>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"